+++ components/firebase/firebase-database.js
@@ -1,5 +1,5 @@
-/*! @license Firebase v3.6.4
-    Build: 3.6.4-rc.2
+/*! @license Firebase v3.6.5
+    Build: 3.6.5-rc.1
     Terms: https://firebase.google.com/terms/
 
     ---